Test instructions added to README.md
