+++ code/firmware/test/Header.hpp
@@ -1,0 +1,7 @@
+#ifndef HEADER_HPP
+#define HEADER_HPP
+
+// Declarations
+int addsometothis();
+
+#endif
