+++ code/firmware/test/README
@@ -1,3 +1,12 @@
+# Instructions for adding tests :
+
+1. Create separate test files for each item with functions
+2. Include above function declarations in Header.hpp
+3. Call the fuctions in test_main.cpp
+
+
+
+
 
 This directory is intended for PlatformIO Test Runner and project tests.
